Removes outline model duplication.
