+++ src/vs/editor/editor.all.ts
@@ -58,6 +58,7 @@ import 'vs/editor/contrib/wordHighlighter/browser/wordHighlighter';
 import 'vs/editor/contrib/wordOperations/browser/wordOperations';
 import 'vs/editor/contrib/wordPartOperations/browser/wordPartOperations';
 import 'vs/editor/contrib/readOnlyMessage/browser/contribution';
+import 'vs/editor/contrib/diffEditorBreadcrumbs/browser/contribution';
 
 // Load up these strings even in VSCode, even if they are not used
 // in order to get them translated
